simulated annealing update
