+++ contribs/common/src/main/java/org/matsim/contrib/common/util/random/RandomUtils.java
@@ -3,7 +3,7 @@
  *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2022 by the members listed in the COPYING,        *
+ * copyright       : (C) 2014 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -17,28 +17,40 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.drt.extension.companions;
+package org.matsim.contrib.common.util.random;
 
-import java.util.List;
-
-import org.matsim.contrib.common.util.random.RandomUtils;
-import org.matsim.contrib.common.util.random.UniformRandom;
-import org.matsim.contrib.common.util.random.WeightedRandomSelection;
+import org.apache.commons.math3.random.*;
 
 /**
- *
- * @author Steffen Axer
- *
+ * Based on org.matsim.core.gbl.MatsimRandom
  */
-public class DrtCompanionUtils {
+public class RandomUtils {
+	public static final int DEFAULT_SEED = 4357;
 
-	public static WeightedRandomSelection<Integer> createIntegerSampler(final List<Double> distribution) {
-		WeightedRandomSelection<Integer> wrs = new WeightedRandomSelection<>(
-				new UniformRandom(RandomUtils.getLocalGenerator()));
-		for (int i = 0; i < distribution.size(); ++i) {
-			wrs.add(i, distribution.get(i));
-		}
-		return wrs;
+	private static final RandomGenerator rg = new MersenneTwister(DEFAULT_SEED);
+
+	private static final UniformRandom uniform = new UniformRandom(rg);
+
+	public static void reset() {
+		reset(DEFAULT_SEED);
 	}
 
+	public static void reset(final int seed) {
+		rg.setSeed(seed);
+	}
+
+	public static RandomGenerator getGlobalGenerator() {
+		return rg;
+	}
+
+	public static UniformRandom getGlobalUniform() {
+		return uniform;
+	}
+
+	/**
+	 * Returns an instance of a random number generator, which can be used locally, e.g. in threads.
+	 */
+	public static RandomGenerator getLocalGenerator() {
+		return new MersenneTwister(rg.nextInt());
+	}
 }
