+++ contribs/drt-extensions/pom.xml
@@ -23,5 +23,11 @@
 			<artifactId>ev</artifactId>
 			<version>15.0-SNAPSHOT</version>
 		</dependency>
+
+		<dependency>
+			<groupId>org.matsim.contrib</groupId>
+			<artifactId>optimization</artifactId>
+			<version>15.0-SNAPSHOT</version>
+		</dependency>
 	</dependencies>
 </project>
