--- contribs/drt-extensions/src/main/java/org/matsim/contrib/optimization/simulatedAnnealing/acceptor/Acceptor.java
@@ -1,11 +1,0 @@
-package org.matsim.contrib.optimization.simulatedAnnealing.acceptor;
-
-import org.matsim.contrib.optimization.simulatedAnnealing.SimulatedAnnealing;
-
-public interface Acceptor<T> {
-
-	boolean accept(SimulatedAnnealing.Solution<T> currentSolution,
-				   SimulatedAnnealing.Solution<T> acceptedSolution,
-				   SimulatedAnnealing.Solution<T> bestSolution
-	);
-}
