--- contribs/drt-extensions/src/main/java/org/matsim/contrib/optimization/simulatedAnnealing/cost/CostCalculator.java
@@ -1,6 +1,0 @@
-package org.matsim.contrib.optimization.simulatedAnnealing.cost;
-
-public interface CostCalculator<T> {
-	double calculateCost(T solution);
-
-}
