--- contribs/drt-extensions/src/main/java/org/matsim/contrib/optimization/simulatedAnnealing/perturbation/Perturbator.java
@@ -1,8 +1,0 @@
-package org.matsim.contrib.optimization.simulatedAnnealing.perturbation;
-
-public interface Perturbator<T> {
-
-	T perturbate(T current);
-
-
-}
