--- contribs/drt-extensions/src/main/java/org/matsim/contrib/optimization/simulatedAnnealing/perturbation/PerturbatorFactory.java
@@ -1,7 +1,0 @@
-package org.matsim.contrib.optimization.simulatedAnnealing.perturbation;
-
-public interface PerturbatorFactory<T> {
-
-	Perturbator<T> createPerturbator();
-
-}
