+++ contribs/dvrp/src/test/java/org/matsim/contrib/util/random/WeightedRandomSelectionTest.java
@@ -24,6 +24,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.apache.commons.lang3.mutable.MutableDouble;
 import org.junit.Before;
 import org.junit.Test;
+import org.matsim.contrib.common.util.random.UniformRandom;
+import org.matsim.contrib.common.util.random.WeightedRandomSelection;
 
 /**
  * @author michalm
