+++ contribs/optimization/src/main/java/simulatedAnnealing/package-info.java
@@ -1,0 +1,20 @@
+/**
+ *
+ * This package contains generic code for iterative optimization using
+ * <a href="https://en.wikipedia.org/wiki/Simulated_annealing">Simulated Annealing</a>.
+ *
+ *
+ * <p>
+ * The idea is that a certain object of the simulation is optimized within the co-evolutionary MATSim
+ * framework (e.g., stop network, schedules, etc...). Similar to the co-evolutionary approach of MATSim,
+ * simulated annealing tries to optimize iteratively by exploring and evaluating new solutions.
+ * In contrast to genetic algorithms, simulated annealing typically work on a single solution at one time.
+ * </p>
+ *
+ *
+ * @since 15.xx
+ * @author nkuehnel
+ */
+package simulatedAnnealing;
+
+
