+++ contribs/pom.xml
@@ -96,7 +96,8 @@
 		<module>sumo</module>
 		<module>taxi</module>
 		<module>vsp</module>
-	</modules>
+        <module>optimization</module>
+    </modules>
 
 	<dependencies>
 		<dependency>
